Fix modal disabling during pending transaction
* added approving & deposting for stake modals

* collect reward hooks added to form components

* withdrawLiquidity hook added to form component

* modal disabling from form components
