+++ src/components/UI/organisms/pools/staking/HarvestForm.jsx
@@ -1,6 +1,8 @@
+import { useEffect } from "react";
 import { RegularButton } from "@/components/UI/atoms/button/regular";
 import AddCircleIcon from "@/icons/AddCircleIcon";
 import { useRegisterToken } from "@/src/hooks/useRegisterToken";
+import { useStakingPoolWithdrawRewards } from "@/src/hooks/useStakingPoolWithdraw";
 import { convertFromUnits } from "@/utils/bn";
 import { formatCurrency } from "@/utils/formatter/currency";
 
@@ -10,10 +12,20 @@ export const HarvestForm = ({
   rewardAmount,
   rewardTokenAddress,
   rewardTokenSymbol,
-  handleWithdrawRewards,
-  withdrawingRewards,
+  poolKey,
+  refetchInfo,
+  setModalDisabled,
 }) => {
+  const { handleWithdrawRewards, withdrawingRewards } =
+    useStakingPoolWithdrawRewards({
+      poolKey,
+      refetchInfo,
+    });
   const { register } = useRegisterToken();
+
+  useEffect(() => {
+    setModalDisabled((val) => ({ ...val, wr: withdrawingRewards }));
+  }, [withdrawingRewards]);
 
   return (
     <div className="px-12">
